refactoring model. Step order column added.

+++ src/main/java/com/test/management/system/repository/StepRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StepRepository extends JpaRepository<Step, Integer> {
+public interface StepRepository extends JpaRepository<Step, Long> {
 }
